app/requirements.txt
cryptography
dash
dash_bootstrap_components
numpy
pandas
plotly
psycopg2-binary

